+++ clang/idxdeclinfo_gen.go
@@ -166,10 +166,10 @@ func (idi IdxDeclInfo) Attributes() []*IdxAttrInfo {
 	return s
 }
 
-func (idi IdxDeclInfo) NumAttributes() uint16 {
-	return uint16(idi.c.numAttributes)
+func (idi IdxDeclInfo) NumAttributes() uint32 {
+	return uint32(idi.c.numAttributes)
 }
 
-func (idi IdxDeclInfo) Flags() uint16 {
-	return uint16(idi.c.flags)
+func (idi IdxDeclInfo) Flags() uint32 {
+	return uint32(idi.c.flags)
 }
